1.4.1
Possibly fix (or atleast catch them) issues with donations.
Fix issue with formatted currency not dropping ".0".

+++ src/main/java/co/casterlabs/koi/util/CurrencyUtil.java
@@ -80,7 +80,7 @@ public class CurrencyUtil {
 
     public static String formatDouble(double value, int fractionalDigits) {
         if (Math.floor(value) == value) {
-            return String.valueOf(Math.floor(value));
+            return String.valueOf((int) Math.floor(value));
         } else {
             return String.format("%." + fractionalDigits + "f", value);
         }
